Sources/requirements.txt
fastapi==0.115.14
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-ollama==0.3.3
motor==3.7.1
pydantic==2.11.7
pymongo==4.13.2
pypdf==5.7.0
pytest==8.4.1
requests==2.32.4
uvicorn[standard]==0.29.0
weaviate-client==3.26.7

# OpenTelemetry dependencies
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asgi
opentelemetry-semantic-conventions
opentelemetry-util-http
opentelemetry-exporter-jaeger
opentelemetry-exporter-jaeger-proto-grpc
opentelemetry-exporter-jaeger-thrift
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
